+++ packages/extension-koni-ui/src/Popup/Confirmations/parts/Detail/Substrate/Extrinsic.tsx
@@ -104,7 +104,7 @@ const mortalityAsString = (era: ExtrinsicEra, hexBlockNumber: string, t: TFuncti
 
 const Component: React.FC<Props> = ({ account, className, payload: { era, nonce, tip }, request: { blockNumber, genesisHash, method, specVersion: hexSpec } }: Props) => {
   const { t } = useTranslation();
-  const chain = useMetadata(genesisHash);
+  const { chain } = useMetadata(genesisHash);
   const chainInfo = useGetChainInfoByGenesisHash(genesisHash);
   const specVersion = useRef(bnToBn(hexSpec)).current;
   const decoded = useMemo(