+++ packages/extension-koni-ui/src/Popup/Confirmations/parts/Sign/Substrate.tsx
@@ -1,27 +1,25 @@
 // Copyright 2019-2022 @subwallet/extension-koni-ui authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-import { ExtrinsicType } from '@subwallet/extension-base/background/KoniTypes';
+import { ExtrinsicType, NotificationType } from '@subwallet/extension-base/background/KoniTypes';
 import { AccountJson, RequestSign } from '@subwallet/extension-base/background/types';
-import { CONFIRMATION_QR_MODAL } from '@subwallet/extension-koni-ui/constants/modal';
+import { _isRuntimeUpdated, detectTranslate, getShortMetadata } from '@subwallet/extension-base/utils';
+import { AlertBox, AlertModal } from '@subwallet/extension-koni-ui/components';
+import { CONFIRMATION_QR_MODAL, NotNeedMigrationGens, SUBSTRATE_GENERIC_KEY } from '@subwallet/extension-koni-ui/constants';
 import { InjectContext } from '@subwallet/extension-koni-ui/contexts/InjectContext';
-import { useGetChainInfoByGenesisHash, useNotification, useParseSubstrateRequestPayload, useUnlockChecker } from '@subwallet/extension-koni-ui/hooks';
-import { useLedger } from '@subwallet/extension-koni-ui/hooks/ledger/useLedger';
+import { useAlert, useGetChainInfoByGenesisHash, useLedger, useMetadata, useNotification, useParseSubstrateRequestPayload, useSelector, useUnlockChecker } from '@subwallet/extension-koni-ui/hooks';
 import { approveSignPasswordV2, approveSignSignature, cancelSignRequest } from '@subwallet/extension-koni-ui/messaging';
-import { RootState } from '@subwallet/extension-koni-ui/stores';
-import { AccountSignMode, PhosphorIcon, SigData, ThemeProps } from '@subwallet/extension-koni-ui/types';
-import { isSubstrateMessage, removeTransactionPersist } from '@subwallet/extension-koni-ui/utils';
-import { getSignMode } from '@subwallet/extension-koni-ui/utils/account/account';
+import { AccountSignMode, PhosphorIcon, SubstrateSigData, ThemeProps } from '@subwallet/extension-koni-ui/types';
+import { getSignMode, isRawPayload, isSubstrateMessage, removeTransactionPersist, toShort } from '@subwallet/extension-koni-ui/utils';
 import { Button, Icon, ModalContext } from '@subwallet/react-ui';
 import CN from 'classnames';
 import { CheckCircle, QrCode, Swatches, Wallet, XCircle } from 'phosphor-react';
 import React, { useCallback, useContext, useEffect, useMemo, useState } from 'react';
-import { useTranslation } from 'react-i18next';
-import { useSelector } from 'react-redux';
+import { Trans, useTranslation } from 'react-i18next';
 import styled from 'styled-components';
 
-import { SignerResult } from '@polkadot/types/types';
-import { SignerPayloadJSON } from '@polkadot/types/types/extrinsic';
+import { SignerPayloadJSON, SignerResult } from '@polkadot/types/types';
+import { hexToU8a, u8aToHex, u8aToU8a } from '@polkadot/util';
 
 import { DisplayPayloadModal, ScanSignature, SubstrateQr } from '../Qr';
 
@@ -31,18 +29,30 @@ interface Props extends ThemeProps {
   request: RequestSign;
   extrinsicType?: ExtrinsicType;
   txExpirationTime?: number;
+  isInternal?: boolean
 }
+
+interface AlertData {
+  description: React.ReactNode;
+  title: string;
+  type: 'info' | 'warning' | 'error';
+}
+const alertModalId = 'dapp-alert-modal';
 
 const handleConfirm = async (id: string) => await approveSignPasswordV2({ id });
 
 const handleCancel = async (id: string) => await cancelSignRequest(id);
 
-const handleSignature = async (id: string, { signature }: SigData) => await approveSignSignature(id, signature);
+const handleSignature = async (id: string, { signature, signedTransaction }: SubstrateSigData) => await approveSignSignature(id, signature, signedTransaction);
 
-const modeCanSignMessage: AccountSignMode[] = [AccountSignMode.QR, AccountSignMode.PASSWORD, AccountSignMode.INJECTED];
+const metadataFAQUrl = 'https://docs.subwallet.app/main/extension-user-guide/faqs#how-do-i-update-network-metadata';
+const genericFAQUrl = 'https://docs.subwallet.app/main/extension-user-guide/faqs#how-do-i-re-attach-a-new-polkadot-account-on-ledger';
+const migrationFAQUrl = 'https://docs.subwallet.app/main/extension-user-guide/faqs#how-do-i-move-assets-from-a-substrate-network-to-the-new-polkadot-account-on-ledger';
+
+const modeCanSignMessage: AccountSignMode[] = [AccountSignMode.QR, AccountSignMode.PASSWORD, AccountSignMode.INJECTED, AccountSignMode.LEGACY_LEDGER, AccountSignMode.GENERIC_LEDGER];
 
 const Component: React.FC<Props> = (props: Props) => {
-  const { account, className, extrinsicType, id, request, txExpirationTime } = props;
+  const { account, className, extrinsicType, id, isInternal, request, txExpirationTime } = props;
 
   const { t } = useTranslation();
   const notify = useNotification();
@@ -50,24 +60,32 @@ const Component: React.FC<Props> = (props: Props) => {
 
   const { activeModal } = useContext(ModalContext);
   const { substrateWallet } = useContext(InjectContext);
+  const { alertProps, closeAlert, openAlert } = useAlert(alertModalId);
 
-  const { chainInfoMap } = useSelector((state: RootState) => state.chainStore);
+  const { chainInfoMap } = useSelector((state) => state.chainStore);
 
-  const payload = useParseSubstrateRequestPayload(request);
+  const genesisHash = useMemo(() => {
+    const _payload = request.payload;
 
-  const [loading, setLoading] = useState(false);
-  const [showQuoteExpired, setShowQuoteExpired] = useState<boolean>(false);
-
+    return isRawPayload(_payload)
+      ? (account.originGenesisHash || chainInfoMap.polkadot.substrateInfo?.genesisHash || '')
+      : _payload.genesisHash;
+  }, [account.originGenesisHash, chainInfoMap.polkadot.substrateInfo?.genesisHash, request.payload]);
   const signMode = useMemo(() => getSignMode(account), [account]);
+  const isLedger = useMemo(() => signMode === AccountSignMode.LEGACY_LEDGER || signMode === AccountSignMode.GENERIC_LEDGER, [signMode]);
 
-  const isLedger = useMemo(() => signMode === AccountSignMode.LEDGER, [signMode]);
+  const { chain, loadingChain } = useMetadata(genesisHash);
+  const chainInfo = useGetChainInfoByGenesisHash(genesisHash);
+  const { addExtraData, hashLoading, isMissingData, payload } = useParseSubstrateRequestPayload(chain, request, isLedger);
+
   const isMessage = isSubstrateMessage(payload);
 
   const approveIcon = useMemo((): PhosphorIcon => {
     switch (signMode) {
       case AccountSignMode.QR:
         return QrCode;
-      case AccountSignMode.LEDGER:
+      case AccountSignMode.LEGACY_LEDGER:
+      case AccountSignMode.GENERIC_LEDGER:
         return Swatches;
       case AccountSignMode.INJECTED:
         return Wallet;
@@ -75,30 +93,173 @@ const Component: React.FC<Props> = (props: Props) => {
         return CheckCircle;
     }
   }, [signMode]);
+  const chainSlug = useMemo(() => signMode === AccountSignMode.GENERIC_LEDGER ? SUBSTRATE_GENERIC_KEY : (chainInfo?.slug || ''), [chainInfo?.slug, signMode]);
+  const networkName = useMemo(() => chainInfo?.name || chain?.name || toShort(genesisHash), [chainInfo, genesisHash, chain]);
+  const isRuntimeUpdated = useMemo(() => {
+    const _payload = request.payload;
 
-  const genesisHash = useMemo(() => {
-    if (isSubstrateMessage(payload)) {
-      return chainInfoMap.polkadot.substrateInfo?.genesisHash || '';
+    if (isRawPayload(_payload)) {
+      return false;
     } else {
-      return payload.genesisHash.toHex();
+      return _isRuntimeUpdated(_payload.signedExtensions);
     }
-  }, [chainInfoMap.polkadot.substrateInfo?.genesisHash, payload]);
+  }, [request.payload]);
+  const requireMetadata = useMemo(() => signMode === AccountSignMode.GENERIC_LEDGER || (signMode === AccountSignMode.LEGACY_LEDGER && isRuntimeUpdated), [isRuntimeUpdated, signMode]);
 
-  const chain = useGetChainInfoByGenesisHash(genesisHash);
+  const isMetadataOutdated = useMemo(() => {
+    const _payload = request.payload;
+
+    if (isRawPayload(_payload)) {
+      return false;
+    } else {
+      const payloadSpecVersion = parseInt(_payload.specVersion);
+      const metadataSpecVersion = chain?.specVersion;
+
+      return payloadSpecVersion !== metadataSpecVersion;
+    }
+  }, [request.payload, chain?.specVersion]);
+
+  const isOpenAlert = !isMessage && !loadingChain && !requireMetadata && !isInternal && (!chain || !chain.hasMetadata || isMetadataOutdated);
+
+  useEffect(() => {
+    if (isOpenAlert) {
+      openAlert({
+        title: t('Pay attention!'),
+        type: NotificationType.WARNING,
+        content: (
+          <Trans
+            components={{
+              highlight: (
+                <a
+                  className='link'
+                  href={metadataFAQUrl}
+                  target='__blank'
+                />
+              )
+            }}
+            i18nKey={detectTranslate("{{networkName}} network's metadata is out of date, which may cause the transaction to fail. Update metadata using <highlight>this guide</highlight> or approve transaction at your own risk")}
+            values={{ networkName }}
+          />),
+        okButton: {
+          text: t('I understand'),
+          icon: CheckCircle,
+          iconWeight: 'fill',
+          onClick: closeAlert
+        }
+      });
+    }
+  }, [closeAlert, isOpenAlert, networkName, openAlert, t]);
+
+  const alertData = useMemo((): AlertData | undefined => {
+    const requireMetadata = signMode === AccountSignMode.GENERIC_LEDGER || (signMode === AccountSignMode.LEGACY_LEDGER && isRuntimeUpdated);
+
+    if (!isMessage && !loadingChain) {
+      if (!chain || !chain.hasMetadata || isMetadataOutdated) {
+        if (requireMetadata) {
+          return {
+            type: 'error',
+            title: t('Error!'),
+            description: (
+              <Trans
+                components={{
+                  highlight: (
+                    <a
+                      className='link'
+                      href={metadataFAQUrl}
+                      target='__blank'
+                    />
+                  )
+                }}
+                i18nKey={detectTranslate("{{networkName}} network's metadata is out of date. Update metadata using <highlight>this guide</highlight> and try again")}
+                values={{ networkName }}
+              />
+            )
+          };
+        }
+      } else {
+        if (isRuntimeUpdated) {
+          if (requireMetadata && isMissingData && !addExtraData) {
+            return {
+              type: 'error',
+              title: t('Error!'),
+              description: t('Unable to sign this transaction on Ledger because the dApp is out of date')
+            };
+          }
+
+          if (signMode === AccountSignMode.LEGACY_LEDGER) {
+            const gens = chain.genesisHash || '___';
+
+            if (NotNeedMigrationGens.includes(gens)) {
+              return {
+                type: 'info',
+                title: t('Helpful tip'),
+                description: (
+                  <Trans
+                    components={{
+                      highlight: (
+                        <a
+                          className='link'
+                          href={genericFAQUrl}
+                          target='__blank'
+                        />
+                      )
+                    }}
+                    i18nKey={detectTranslate('To sign this transaction, open “Polkadot” app on Ledger, hit Refresh and Approve again. For a better experience, re-attach your Polkadot new account using <highlight>this guide</highlight>')}
+                  />
+                )
+              };
+            } else {
+              return {
+                type: 'info',
+                title: t('Helpful tip'),
+                description: (
+                  <Trans
+                    components={{
+                      highlight: (
+                        <a
+                          className='link'
+                          href={migrationFAQUrl}
+                          target='__blank'
+                        />
+                      )
+                    }}
+                    i18nKey={detectTranslate('To sign this transaction, open “Polkadot Migration” app on Ledger, hit Refresh and Approve again. For a better experience, move your assets on {{networkName}} network to the Polkadot new account using <highlight>this guide</highlight>')}
+                    values={{ networkName }}
+                  />
+                )
+              };
+            }
+          }
+        } else {
+          if (signMode === AccountSignMode.GENERIC_LEDGER) {
+            return {
+              type: 'error',
+              title: t('Error!'),
+              description: t('Unable to sign this transaction on Ledger because the {{networkName}} network is out of date', { replace: { networkName } })
+            };
+          }
+        }
+      }
+    }
+
+    return undefined;
+  }, [addExtraData, chain, isMessage, isMetadataOutdated, isMissingData, isRuntimeUpdated, loadingChain, networkName, signMode, t]);
+
+  const activeLedger = useMemo(() => isLedger && !loadingChain && alertData?.type !== 'error', [isLedger, loadingChain, alertData?.type]);
 
   const { error: ledgerError,
     isLoading: isLedgerLoading,
     isLocked,
     ledger,
     refresh: refreshLedger,
-    signTransaction: ledgerSign,
-    warning: ledgerWarning } = useLedger(chain?.slug, isLedger);
+    signMessage: ledgerSignMessage,
+    signTransaction: ledgerSignTransaction,
+    warning: ledgerWarning } = useLedger(chainSlug, activeLedger, true, isRuntimeUpdated || isMessage);
 
-  const isLedgerConnected = useMemo(() => !isLocked && !isLedgerLoading && !!ledger, [
-    isLedgerLoading,
-    isLocked,
-    ledger
-  ]);
+  const isLedgerConnected = useMemo(() => !isLocked && !isLedgerLoading && !!ledger, [isLedgerLoading, isLocked, ledger]);
+
+  const [loading, setLoading] = useState(false);
+  const [showQuoteExpired, setShowQuoteExpired] = useState<boolean>(false);
 
   // Handle buttons actions
   const onCancel = useCallback(() => {
@@ -122,7 +283,7 @@ const Component: React.FC<Props> = (props: Props) => {
     }, 1000);
   }, [id]);
 
-  const onApproveSignature = useCallback((signature: SigData) => {
+  const onApproveSignature = useCallback((signature: SubstrateSigData) => {
     setLoading(true);
 
     setTimeout(() => {
@@ -141,7 +302,7 @@ const Component: React.FC<Props> = (props: Props) => {
   }, [activeModal]);
 
   const onConfirmLedger = useCallback(() => {
-    if (!payload || typeof payload === 'string') {
+    if (!payload) {
       return;
     }
 
@@ -153,28 +314,66 @@ const Component: React.FC<Props> = (props: Props) => {
 
     setLoading(true);
 
-    setTimeout(() => {
-      const payloadU8a = payload.toU8a(true);
+    // eslint-disable-next-line @typescript-eslint/no-misused-promises
+    setTimeout(async () => {
+      if (typeof payload === 'string') {
+        try {
+          const { signature } = await ledgerSignMessage(u8aToU8a(payload), account.accountIndex, account.addressOffset);
 
-      ledgerSign(payloadU8a, account.accountIndex, account.addressOffset)
-        .then(({ signature }) => {
           onApproveSignature({ signature });
-        })
-        .catch((e: Error) => {
-          console.log(e);
-          setLoading(false);
-        });
-    });
-  }, [
-    account.accountIndex,
-    account.addressOffset,
-    isLedgerConnected,
-    ledger,
-    ledgerSign,
-    onApproveSignature,
-    payload,
-    refreshLedger
-  ]);
+        } catch (e) {
+          console.error(e);
+        }
+
+        setLoading(false);
+      } else {
+        const payloadU8a = payload.toU8a(true);
+
+        let metadata: Uint8Array;
+
+        if (isRuntimeUpdated) {
+          try {
+            const blob = u8aToHex(payloadU8a);
+            const shortener = await getShortMetadata(chainInfo?.slug || '', blob);
+
+            metadata = hexToU8a(shortener);
+          } catch (e) {
+            notify({
+              message: (e as Error).message,
+              type: 'error'
+            });
+            setLoading(false);
+
+            return;
+          }
+        } else {
+          metadata = new Uint8Array(0);
+        }
+
+        try {
+          const { signature } = await ledgerSignTransaction(payloadU8a, metadata, account.accountIndex, account.addressOffset);
+
+          if (addExtraData) {
+            const extrinsic = payload.registry.createType(
+              'Extrinsic',
+              { method: payload.method },
+              { version: 4 }
+            );
+
+            extrinsic.addSignature(account.address, signature, payload.toHex());
+
+            onApproveSignature({ signature, signedTransaction: extrinsic.toHex() });
+          } else {
+            onApproveSignature({ signature });
+          }
+        } catch (e) {
+          console.error(e);
+        }
+
+        setLoading(false);
+      }
+    }, 100);
+  }, [account, chainInfo, isLedgerConnected, isRuntimeUpdated, ledger, ledgerSignMessage, ledgerSignTransaction, addExtraData, notify, onApproveSignature, payload, refreshLedger]);
 
   const onConfirmInject = useCallback(() => {
     if (substrateWallet) {
@@ -196,8 +395,14 @@ const Component: React.FC<Props> = (props: Props) => {
 
       setLoading(true);
       promise
-        .then(({ signature }) => {
-          onApproveSignature({ signature });
+        .then(({ signature, signedTransaction: _signedTransaction }) => {
+          const signedTransaction = _signedTransaction
+            ? _signedTransaction instanceof Uint8Array
+              ? u8aToHex(_signedTransaction)
+              : _signedTransaction
+            : undefined;
+
+          onApproveSignature({ signature, signedTransaction });
         })
         .catch((e) => {
           console.error(e);
@@ -227,7 +432,8 @@ const Component: React.FC<Props> = (props: Props) => {
       case AccountSignMode.QR:
         onConfirmQr();
         break;
-      case AccountSignMode.LEDGER:
+      case AccountSignMode.LEGACY_LEDGER:
+      case AccountSignMode.GENERIC_LEDGER:
         onConfirmLedger();
         break;
       case AccountSignMode.INJECTED:
@@ -274,57 +480,82 @@ const Component: React.FC<Props> = (props: Props) => {
   }, [txExpirationTime]);
 
   return (
-    <div className={CN(className, 'confirmation-footer')}>
-      <Button
-        disabled={loading}
-        icon={(
-          <Icon
-            phosphorIcon={XCircle}
-            weight='fill'
-          />
-        )}
-        onClick={onCancel}
-        schema={'secondary'}
-      >
-        {t('Cancel')}
-      </Button>
-      <Button
-        disabled={showQuoteExpired || (isMessage && !modeCanSignMessage.includes(signMode))}
-        icon={(
-          <Icon
-            phosphorIcon={approveIcon}
-            weight='fill'
-          />
-        )}
-        loading={loading}
-        onClick={onConfirm}
-      >
-        {
-          signMode !== AccountSignMode.LEDGER
-            ? t('Approve')
-            : !isLedgerConnected
-              ? t('Refresh')
-              : t('Approve')
-        }
-      </Button>
+    <>
       {
-        signMode === AccountSignMode.QR && (
-          <DisplayPayloadModal>
-            <SubstrateQr
-              address={account.address}
-              genesisHash={genesisHash}
-              payload={payload || ''}
-            />
-          </DisplayPayloadModal>
+        alertData && (
+          <AlertBox
+            className={CN(className, 'alert-box')}
+            description={alertData.description}
+            title={alertData.title}
+            type={alertData.type}
+          />
         )
       }
-      {signMode === AccountSignMode.QR && <ScanSignature onSignature={onApproveSignature} />}
-    </div>
+      {
+        !!alertProps && (
+          <AlertModal
+            modalId={alertModalId}
+            {...alertProps}
+          />
+        )
+      }
+      <div className={CN(className, 'confirmation-footer')}>
+        <Button
+          disabled={loading}
+          icon={(
+            <Icon
+              phosphorIcon={XCircle}
+              weight='fill'
+            />
+          )}
+          onClick={onCancel}
+          schema={'secondary'}
+        >
+          {t('Cancel')}
+        </Button>
+        <Button
+          disabled={showQuoteExpired || loadingChain || hashLoading || (isMessage ? !modeCanSignMessage.includes(signMode) : alertData?.type === 'error')}
+          icon={(
+            <Icon
+              phosphorIcon={approveIcon}
+              weight='fill'
+            />
+          )}
+          loading={loading}
+          onClick={onConfirm}
+        >
+          {
+            !isLedger
+              ? t('Approve')
+              : !isLedgerConnected
+                ? t('Refresh')
+                : t('Approve')
+          }
+        </Button>
+        {
+          signMode === AccountSignMode.QR && (
+            <DisplayPayloadModal>
+              <SubstrateQr
+                address={account.address}
+                genesisHash={genesisHash}
+                payload={payload || ''}
+              />
+            </DisplayPayloadModal>
+          )
+        }
+        {signMode === AccountSignMode.QR && <ScanSignature onSignature={onApproveSignature} />}
+      </div>
+    </>
   );
 };
 
 const SubstrateSignArea = styled(Component)<Props>(({ theme: { token } }: Props) => {
-  return {};
+  return {
+    '&.alert-box': {
+      margin: token.padding,
+      marginBottom: 0
+    }
+  };
 });
 
 export default SubstrateSignArea;